+++ src/binaryTranslator/main.cpp
@@ -47,7 +47,7 @@ void finish (List<basicblock*> *bbList, List<basicblock*> *phBBList, std::string
 int main (int argc, char* argv[])
 {
 	Assert (argc == 4 && "USAGE: ./PhraseFormer <program_name> <reg_alloc_method> <scheduling_method>");
-	unsigned t0=clock (),t1;
+	long long unsigned t0 = clock (), t1;
 	//SETUP VARIABLES
 	List<instruction*>* insList = new List<instruction*>;
 	List<basicblock*>* bbList   = new List<basicblock*>;
@@ -118,7 +118,7 @@ int main (int argc, char* argv[])
 	/* ---------------------*/
 	printf ("- Make Dot Files & Stat Data -\n");
 	finish (bbList, phBBList, &program_name, sch_mode, reg_alloc_mode);
-	t1=clock ()-t0;
-	printf ("\n-------------\nEXECUTION COMPLETED SUCCESSFULLY. (Time: %f Minutes)\n-------------\n", ( (float)t1)/CLOCKS_PER_SEC/60.0);
+	t1 = clock () - t0;
+	printf ("\n-------------\nEXECUTION COMPLETED SUCCESSFULLY. (Time: %f Minutes)\n-------------\n", (double)(t1 / CLOCKS_PER_SEC) / 60.0);
 	return 0;
 }
